+++ src/main/kotlin/com/cluttered/cryptocurrency/model/MarketHistory.kt
@@ -1,14 +1,15 @@
 package com.cluttered.cryptocurrency.model
 
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 import java.time.ZonedDateTime
 
 data class MarketHistory(
         @SerializedName("Id") val id: Long,
         @SerializedName("TimeStamp") val timestamp: ZonedDateTime,
-        @SerializedName("Quantity") val quantity: Double,
-        @SerializedName("Price") val price: Double,
-        @SerializedName("Total") val total: Double,
+        @SerializedName("Quantity") val quantity: BigDecimal,
+        @SerializedName("Price") val price: BigDecimal,
+        @SerializedName("Total") val total: BigDecimal,
         @SerializedName("FillType") val fillType: FillType,
         @SerializedName("OrderType") val orderType: OrderType
 ) {
