+++ src/main/kotlin/com/cluttered/cryptocurrency/model/OpenOrder.kt
@@ -1,6 +1,7 @@
 package com.cluttered.cryptocurrency.model
 
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 import java.time.ZonedDateTime
 import java.util.*
 
@@ -8,12 +9,12 @@ data class OpenOrder(
         @SerializedName("OrderUuid") val orderUuid: UUID,
         @SerializedName("Exchange") val exchange: String,
         @SerializedName("OrderType") val orderType: OrderType,
-        @SerializedName("Quantity") val quantity: Double,
-        @SerializedName("QuantityRemaining") val quantityRemaining: Double,
-        @SerializedName("Limit") val limit: Double,
-        @SerializedName("CommissionPaid") val commissionPaid: Double,
-        @SerializedName("Price") val price: Double,
-        @SerializedName("PricePerUnit") val pricePerUnit: Double,
+        @SerializedName("Quantity") val quantity: BigDecimal,
+        @SerializedName("QuantityRemaining") val quantityRemaining: BigDecimal,
+        @SerializedName("Limit") val limit: BigDecimal,
+        @SerializedName("CommissionPaid") val commissionPaid: BigDecimal,
+        @SerializedName("Price") val price: BigDecimal,
+        @SerializedName("PricePerUnit") val pricePerUnit: BigDecimal,
         @SerializedName("Opened") val opened: ZonedDateTime,
         @SerializedName("Closed") val closed: ZonedDateTime?,
         @SerializedName("CancelInitiated") val cancelInitiated: Boolean,
