+++ src/main/kotlin/com/cluttered/cryptocurrency/model/Ticker.kt
@@ -1,9 +1,10 @@
 package com.cluttered.cryptocurrency.model
 
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 data class Ticker(
-        @SerializedName("Bid") val bid: Double,
-        @SerializedName("Ask") val ask: Double,
-        @SerializedName("Last") val last: Double
+        @SerializedName("Bid") val bid: BigDecimal,
+        @SerializedName("Ask") val ask: BigDecimal,
+        @SerializedName("Last") val last: BigDecimal
 )
